test: set ReplicaSetSetupTimeout for every embedded MongoDB
WatchToGrpcTests.UseMongoForResultsShouldSucceed failed intermittently on the
Windows runner with "Replica set initialization command took longer than the
specified timeout of 10 seconds": starting a single-node replica set means
launching mongod, running rs.initiate and waiting for PRIMARY, which does not
fit in EphemeralMongo's 10s default on a loaded runner that also has to
download mongod first.

MongoDatabaseProvider has carried a 30s timeout since 2023 and does not flake,
but the MongoRunnerOptions block was copied into all five test providers, so
the setting never reached the other four. Move that block into a shared
MongoRunnerOptionsFactory and route every provider through it, so the next
option cannot be stranded in one file either.
